backend/requirements.txt
Flask>=2.2.5
Jinja2>=3.0.3
gunicorn==21.2.0
flask-cors==4.0.0
Flask_Migrate==4.1.0
Flask_SQLAlchemy==3.1.0
passlib==1.7.4
PyJWT==2.10.1
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.32.3
Werkzeug>=2.2.0 
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
click>=8.0.0
Flask-CLI==0.4.0
faker==18.13.0